requirements.txt
django
djangorestframework
pillow
djangorestframework-simplejwt
